+++ transtool/config/config.py
@@ -45,8 +45,7 @@ class Config(object):
         self.quiet: bool = False
         self.verbose: bool = False
 
-        self.update: bool = False
-        self.create: bool = False
+        self.write: bool = False
         self.write_reference: bool = False
 
         self.files: List[str] = []